Encapsulte SharedPreference on its own class
